// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/relatedAnswer.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type RelatedAnswer = 0 | 1 | 2 | 3 | 4 | 5 | 6;

export const RelatedAnswer = {
    NUMBER_0: 0 as RelatedAnswer,
    NUMBER_1: 1 as RelatedAnswer,
    NUMBER_2: 2 as RelatedAnswer,
    NUMBER_3: 3 as RelatedAnswer,
    NUMBER_4: 4 as RelatedAnswer,
    NUMBER_5: 5 as RelatedAnswer,
    NUMBER_6: 6 as RelatedAnswer
};