// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/checkboxQuestionAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseModelAC } from './baseModelAC';
import { OptionAC } from './optionAC';
import { RelatedAnswer } from './relatedAnswer';

export interface CheckboxQuestionAC extends BaseModelAC { 
    isOtherToBeShown: boolean;
    relatedAnswer: RelatedAnswer;
    fieldLabel?: string;
    options?: Array<OptionAC>;
}